Fixed installation to work with mac os x !

+++ install.sh
@@ -49,11 +49,17 @@ cp vcftools_0.1.11/bin/* bin/
 cp vcftools_0.1.11/perl/*pm lib/perl5/
 rm -Rf vcftools_0.1.11
 echo "Installing QCTool"
-tar xzf src/qctool_v1.3-linux-x86_64.tgz
-mv qctool_v1.3-linux-x86_64/qctool bin/
-rm -Rf qctool_v1.3-linux-x86_64
-echo "Installing Shapeit"
+if [ `uname` = "Darwin" ]; then 
+    tar xzf src/qctool_v1.4-osx.tgz
+    mv qctool_v1.4-osx/qctool bin/
+    rm -Rf qctool_v1.4-osx
+else
+    tar xzf src/qctool_v1.4-linux-x86_64.tgz
+    mv qctool_v1.4-linux-x86_64/qctool bin/
+    rm -Rf qctool_v1.4-linux-x86_64
+fi
 
+echo "Installing Shapeit"
 if [ `uname` = "Darwin" ]; then
     wget https://mathgen.stats.ox.ac.uk/genetics_software/shapeit/shapeit.v2.r778.MacOSX.tgz
     tar xzf shapeit.v2.r778.MacOSX.tgz
@@ -90,13 +96,17 @@ cp tabix-0.2.6/bgzip bin/
 cp tabix-0.2.6/tabix bin/
 rm -Rf tabix-0.2.6
 echo "Installing Variscan"
-tar -xzf src/variscan-2.0.3.tar.gz
-(cd variscan-2.0.3/src/ && rm *o)
-change_folder  variscan-2.0.3
-check_success bash autogen.sh && make
-orig_dir
-mv variscan-2.0.3/src/variscan bin/
-rm -Rf variscan-2.0.3
+if [ `uname` = "Darwin" ]; then
+    echo "Cannot install on OSX"
+else
+    tar -xzf src/variscan-2.0.3.tar.gz
+    (cd variscan-2.0.3/src/ && rm *o)
+    change_folder  variscan-2.0.3
+    check_success bash autogen.sh && make
+    orig_dir
+    mv variscan-2.0.3/src/variscan bin/
+    rm -Rf variscan-2.0.3
+fi
 echo "Installing Beagle"
 cp src/beagle.jar bin/
 echo "Installing getopt"
